requirements.txt
langchain
langchain-community
requests
python-dotenv
langchain_groq
pyttx3